Better error handling for ill-formatted messages
Before this, on android, if there was a single ill-formatted message, we used
to abort the entire sync. Now, we only skip that message. Currently, all
skipped messages are deleted at the end of the sync.

We were originally going to store skipped messages in a separate error
database. But it turns out that on iOS, we check validity at insert time, and
invalid messages are not stored in the usercache at all, which means that they
are effectively discarded.

We need to decide our error handling policy. Until we do this, we discard on
both iOS (at insert time) and on android (at upload time). We have created an
error database to store the data, but are not using it pending decision.

The issue is being tracked in:
https://github.com/e-mission/cordova-usercache/issues/7