Fix deletion of RW-documents + robustify phone_to_server
Fix the deletion of the RW documents that have been superceded. This needs to
be done BEFORE the general clear - otherwise, the superceding document will be
deleted and we will retain the rw-document forever.

Also, reading all entries from the phone into memory can cause out of memory
issues.  Push data in smaller batches instead, which is not a long-term
solution, but is a reasonable alternative until we come up with a streaming
solution.
https://github.com/e-mission/e-mission-data-collection/issues/109