Unify the access to the usercache through the native plugin
This fixes #124.

Before this, we used to access the usercache in native code using the plugin,
and in javascript using a separate sqlite plugin that accessed the same table.
Unfortunately, accessing the same table through two different handlers runs into locking issues.
We unify all access through the usercache plugin to avoid that. It also avoids leaking information
such as the table name and table column names outside the native code and make this a lot more modular.

This may also be the cause for #126, and thus, for assorted battery issues that
have been reported on iOS.