Final set of changes
We sometimes have a document and sometimes have an array. For the trips, we
have a JSON array.  For the consent, we have a document.

We can't know what it is until we parse it.

So we try one first, catch the exception and try the other one.
Both the diary and the consent work with this.