Extend the unified usercache information to return both data and metadata
The original, non-unified version of the usercache returned the metadata in
addition to the data, so that we could list show the raw sensor data, etc.
When we switched to the unified version, we switched to the UserCache interface
which only returned data.

It seems like it would be useful to have the ability to return both data and
metadata depending on our needs. So I add a `withMetadata` flag for all the get
methods.

This required a bunch of restructuring in the native code, but nothing too
complicated. Just basically fairly mechanical, straightforward changes.

Not been tested yet. Will merge after testing.