Delete obsolete documents from the usercache + split clean function
This fixes #24. We do this by deleting all documents, not just the one in the
specified timequery. We will get new documents from the cache anyway, so this
is not a big deal.

However, the naive addition of this code to clearEntries introduced
https://github.com/e-mission/cordova-jwt-auth/issues/8

This is because on iOS, before this change:
- we push and pull the data in parallel
- we clear after the push

So if the push happens to finish before the pull, we will end up with 0 entries
in the cache. This means that isConsented will return false, and we can't even
sync any more.

The fix for this is to split the `clearEntries` into multiple functions, and call them
correctly - entries and rw-docs for push, and documents for pull.

And while debugging the deletion code, I discovered that the fix for #13 was
that `rawQuery` only works for actual queries (e.g. SELECT statements). For
DELETE, you need to use `db.delete` instead and that appears to work.