Fix minor potential issue with clearing rw-documents
Before this change, we deleted all rw documents except the most recent one.
However, we are not guaranteed to have pushed everything upto the most recent
one.

In particular, let us assume that a sync starts at time t_s.
All entries upto t_s are read and we start pushing them to the server.
The user changes the config at t_1 (while sync is ongoing)
The user changes the config at t_2 (while sync is ongoing)
The user changes the config at t_3 (while sync is ongoing)
Sync completes and deletes configs at t_1 and t_2 although they have not been pushed to the server.

This is not a big issue, specially for the config because very quick configs
are not a great use case anyway. but we expect to have the history, and there
may be rw-documents that we expect that behavior for later.

And the fix is really simple - just delete upto MIN(endTs, max(write_ts).
So in this case, we would delete everything before t_s, and all subsequent
changes will be pushed to the server.