Switch all the user stats to the user cache
Back when we stored the raw data directly as sections and trips, we used to
store the stats in a timeseries. Now that we are storing the raw data in a
timeseries, and are planning to use the stats for analysis into user behavior,
it would be more convenient to store the stats into the same timeseries. Then,
the timeseries is really the a full representation of the source of truth, we
can reuse similar code for analysis.

This change has several components.
- In the usercache, since we already have the functionality, we just need to add a new set of keys.
  We also need to move the app_stats keys to this repo since we are going to
  delete the entire old client stats repo. Woo!!
- Add a new wrapper for a stats event since we need to send that to the server
- Currently, the only client stats that we report from native code are in the
  sync code path. Change all of those to use usercache stats.
- Move the app_launched stat to the data collection plugin load
- From the `CommunicationHelper`, remove all instances of the `setStats` call

The related server commit is https://github.com/e-mission/e-mission-server/commit/7487c82578e8933f4da8f9d3fa3522c102906c81
from pull request https://github.com/e-mission/e-mission-server/pull/443