:sparkles: New methods to return all keys and all local storage keys
The list of all keys is just for debugging. It allows us to answer the
question: "when these local storage entries are deleted, is everything
deleted? Or is there something special/unique about local storage?"

It would be pretty terrible if we are losing actual background information as
well, since we would be losing trip data as well.

The list of local storage keys will actually be used in the phone code for
working around the issue that the SQLite entries of the local storage type are
sometimes deleted.
https://github.com/e-mission/e-mission-docs/issues/930

We already had an issue in which the webview based local storage was getting
arbitrarily deleted.

Since either of the entries can be deleted, we treat don't treat one as primary
and the other as secondary. Instead, we treat them as co-equal, duplicate
entries (similar to RAID :). We will use the keys to determine which ones are
missing from each of the copies and copy them over properly.